Added items removal

+++ resources/views.py
@@ -12,7 +12,7 @@ from django.contrib.auth.models import Group
 from django.views.generic import TemplateView
 
 from .models import Resource, Item, Storage
-from .forms import ResourceForm, ItemForm, DelItemForm, GroupForm, DeleteResourceForm
+from .forms import ResourceForm, ItemForm, DeleteItemForm, GroupForm, DeleteResourceForm
 from .encryption import symEncrypt_b64, symDecrypt_b64
 
 
@@ -132,7 +132,6 @@ def detail(request, resource_id):
             item = Item(resource_id=resource_id)
             item.key = symEncrypt_b64(sym_key, form.cleaned_data["item_key"])
             item.val = symEncrypt_b64(sym_key, form.cleaned_data["item_val"])
-            item.url = symEncrypt_b64(sym_key, form.cleaned_data["item_url"])
             item.save()
             return HttpResponseRedirect("/resources/{}/".format(resource_id))
     else:
@@ -147,12 +146,13 @@ def decryptItem(key_b64, item):
         'pk': item.id,
         'key': symDecrypt_b64(key_b64, item.key),
         'val': symDecrypt_b64(key_b64, item.val),
-        'url': symDecrypt_b64(key_b64, item.url),
     }
 
 
 def delete_resource(request, resource_id):
     resource = get_object_or_404(Resource, pk=resource_id)
+    if not request.user.groups.filter(pk=resource.group_id).exists():
+        raise Http404
     group_id = resource.group_id
     if request.method == 'POST':
         form = DeleteResourceForm(request.POST)
@@ -164,5 +164,18 @@ def delete_resource(request, resource_id):
     return render(request, "resources/delete.html", context)
 
 
-def delete_item(request):
-    pass
+def delete_item(request, item_id):
+    item = get_object_or_404(Item, pk=item_id)
+    resource = item.resource
+    if not request.user.groups.filter(pk=resource.group_id).exists():
+        raise Http404
+    if request.method == 'POST':
+        form = DeleteItemForm(request.POST)
+        if form.is_valid():
+            item.delete()
+            return HttpResponseRedirect("/resources/{}".format(resource.pk))
+    form = DeleteItemForm()
+    context = {"item": item, "form": form}
+    return render(request, "items/delete.html", context)
+
+
